lib/core: Move current voltage monitor API from hwtest
This commit moves APIs related to current voltage monitor from
test/hwtest to lib.
