+++ lib/core/CMakeLists.txt
@@ -6,7 +6,9 @@ if (CONFIG_BOARD_SCSAT1_MAIN OR CONFIG_BOARD_SCSAT1_ADCS)
 	zephyr_library_sources(wdog.c)
 	zephyr_library_sources_ifdef(CONFIG_SC_LIB_CORE_MAIN pwrctrl_main.c)
 	zephyr_library_sources_ifdef(CONFIG_SC_LIB_CORE_MAIN temp_main.c)
+	zephyr_library_sources_ifdef(CONFIG_SC_LIB_CORE_MAIN cv_main.c)
 	zephyr_library_sources_ifdef(CONFIG_SC_LIB_CORE_ADCS pwrctrl_adcs.c)
 	zephyr_library_sources_ifdef(CONFIG_SC_LIB_CORE_ADCS temp_adcs.c)
+	zephyr_library_sources_ifdef(CONFIG_SC_LIB_CORE_ADCS cv_adcs.c)
 	zephyr_include_directories(.)
 endif()
