+++ lib/core/cv_adcs.c
@@ -5,10 +5,10 @@
  */
 
 #include <zephyr/drivers/i2c.h>
-#include "cv.h"
+#include "cv_adcs.h"
 
 #include <zephyr/logging/log.h>
-LOG_MODULE_REGISTER(cv, CONFIG_SCSAT1_ADCS_LOG_LEVEL);
+LOG_MODULE_REGISTER(cv, CONFIG_SC_LIB_CORE_ADCS_LOG_LEVEL);
 
 #define CV_ADCS_ADDR (0x42)
 #define CV_RW_ADDR   (0x40)
