+++ lib/core/cv_main.h
