+++ tests/hwtest/adcs/CMakeLists.txt
@@ -10,7 +10,6 @@ project(scsat1-adcs)
 
 target_sources(app PRIVATE src/main.c)
 target_sources(app PRIVATE src/temp_test.c)
-target_sources(app PRIVATE src/cv.c)
 target_sources(app PRIVATE src/cv_test.c)
 target_sources(app PRIVATE src/imu.c)
 target_sources(app PRIVATE src/imu_test.c)
