+++ tests/hwtest/adcs/src/cv_test.c
@@ -6,7 +6,7 @@
 
 #include <zephyr/kernel.h>
 #include "common.h"
-#include "cv.h"
+#include "cv_adcs.h"
 #include "cv_test.h"
 #include "sc_fpgamon.h"
