+++ tests/hwtest/main/CMakeLists.txt
@@ -12,7 +12,6 @@ target_sources(app PRIVATE src/main.c)
 target_sources(app PRIVATE src/mgnm.c )
 target_sources(app PRIVATE src/mgnm_test.c )
 target_sources(app PRIVATE src/temp_test.c)
-target_sources(app PRIVATE src/cv.c)
 target_sources(app PRIVATE src/cv_test.c)
 target_sources(app PRIVATE src/csp.c)
 target_sources(app PRIVATE src/csp_test.c)
